Build Windows tools using gyp as well
dump_syms is a big problem, because it requires the MS DIA (Debug Interface Access) SDK.
Note that the DIA SDK is not in the Express editions of Visual Studio.

DIA SDK uses ATL.

See https://groups.google.com/forum/#!topic/google-breakpad-discuss/F0jMWxmWk0M
for a change to remove that dependency.